QByteArray: in/deflate: compare different types as size_t
Compiling for android a certain configuration warns about comparisons
between types of different signedness.
On 32-bit we cannot cast the unsigned type to qsizetype and on x64
we cannot cast the qsizetype to Zlib's type (both potentially truncating).
So, cast both to size_t before comparing

Pick-to: 6.4 6.2

+++ src/corelib/text/qbytearray.cpp
@@ -620,7 +620,8 @@ static QByteArray xxflate(ZLibOp op, QArrayDataPointer<char> out, QByteArrayView
                 avail_out = capacity - out.size;
             }
             zs.next_out = reinterpret_cast<uchar *>(out.data()) + out.size;
-            zs.avail_out = avail_out > MaxChunkSize ? MaxChunkSize : ZlibChunkSize_t(avail_out);
+            zs.avail_out = size_t(avail_out) > size_t(MaxChunkSize) ? MaxChunkSize
+                                                                    : ZlibChunkSize_t(avail_out);
             out.size += zs.avail_out;
 
             Q_ASSERT(zs.avail_out > 0);
